extract to separate component and clean up ux
