+++ src/components/civ-pool-container/index.tsx
@@ -1,35 +1,20 @@
-import { ChangeEvent, FC, useState } from 'react';
+import { FC, useMemo } from 'react';
 
-import './civ-pool-container.scss';
-import { CivPool } from '../civ-pool';
-import { CivPoolButtonBar } from '../civ-pool-button-bar';
 import { useAppSelector } from '../../hooks';
 import { selectCivs } from '../../store/slices/civs-slice';
-import { selectDrafts } from '../../store/slices/drafts-slice';
+import { CivPool } from '../civ-pool';
+import { CivPoolButtonBar } from '../civ-pool-button-bar';
+import { Draft } from '../draft';
+import './civ-pool-container.scss';
 
 export interface ICivPoolContainerProps {}
 
 export const CivPoolContainer: FC<ICivPoolContainerProps> = (props) => {
   const { allCivs, civPool } = useAppSelector(selectCivs);
-  const { draft } = useAppSelector(selectDrafts);
-  const civPoolCount = civPool.length > 0 ? civPool.length : allCivs.length;
-
-  const [editName, setEditName] = useState(draft?.name || '');
-  const [editDescription, setEditDescription] = useState(draft?.desc || '');
-
-  const handleEditName = (element: ChangeEvent<HTMLInputElement>) => {
-    const newName = element.target.value;
-    if (newName.length < 51) {
-      setEditName(newName);
-    }
-  };
-
-  const handleEditDescription = (element: ChangeEvent<HTMLInputElement>) => {
-    const newDescription = element.target.value;
-    if (newDescription.length < 127) {
-      setEditDescription(newDescription);
-    }
-  };
+  const civPoolCount = useMemo(
+    () => (civPool.length > 0 ? civPool.length : allCivs.length),
+    [civPool],
+  );
 
   return (
     <>
@@ -42,20 +27,7 @@ export const CivPoolContainer: FC<ICivPoolContainerProps> = (props) => {
         {civPoolCount !== 1 && 's'} in current draft pool
       </p>
       <div className='civ-pool-container'>
-        {draft && (
-          <div className='draft-name-container'>
-            <input
-              className='draft-name'
-              value={editName}
-              onChange={handleEditName}
-            />
-            <input
-              className='draft-description'
-              value={editDescription}
-              onChange={handleEditDescription}
-            />
-          </div>
-        )}
+        <Draft />
         <CivPool />
         <CivPoolButtonBar />
       </div>
